requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosmtplib==2.0.1
alembic==1.10.4
anyio==3.6.2
async-exit-stack==1.0.1
async-generator==1.10
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
databases==0.7.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.95.1
fastapi-mail==1.2.8
frozenlist==1.3.3
h11==0.14.0
httplib2==0.22.0
idna==3.4
install==1.3.5
Jinja2==3.1.2
jose==1.0.0
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
oauth2==1.9.0.post1
openai==0.27.6
passlib==1.7.4
psycopg2==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
PyMuPDF==1.22.2
pyparsing==3.0.9
PyPDF2==3.0.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
yarl==1.9.1
